タイトル画面とスコアゲージの作成

+++ SWICTH_Prototype/Assets/Scripts/Score.cs
@@ -8,7 +8,7 @@ public class Score : MonoBehaviour
 {
 
     [SerializeField] private TextMeshProUGUI scoreText;
-    static public int score;
+    static public float score;
 
     // Start is called before the first frame update
     void Start()
